One of those fixes isn't necessary anymore

+++ PlanningPoker/src/edu/wpi/cs/wpisuitetng/modules/planningpoker/controller/GameTimeoutObserver.java
@@ -71,7 +71,7 @@ public class GameTimeoutObserver extends Thread {
      * @param game
      */
     public GameTimeoutObserver(Session session, GameModel game) {
-        super("GameTimeoutObserver-" + (long)game.getID());
+        super("GameTimeoutObserver-" + game.getID());
         dateFormatGMT.setTimeZone(TimeZone.getTimeZone("GMT"));
         
         this.session = session;
